text-support mixin should use on('didInsertElement')
this let's users of the mixin implement their on version

I took the chance to add a simple test for this use case.
It turns out it's probably a good idea for us to export also the
templates as I'd rather test against the real templates than using a
fake version.

Thoughts? /cc @elucid @jacojoubert